// src/app/shared/models/racketDto.ts
export interface racketDto {
  racketId: number
  prezzo: number
  vecchioPrezzo: number
  marca: string
  modello: string
  sesso: string
  imageLink: string
  coloreUno: string
  coloreDue: string
  profilo: number
  lunghezza: number
  peso: string
  numeroArticolo: string
  puntoDiEquilibrio: number
  tipoDiGioco: string
  url: string
  tipoDiProdotto: string
  telaio: string
  nucleo: string
  livelloDiGioco: string
  forma: string
  eta: number
  bilanciamento: string
  anno: number
}